backend/requirements.txt
flask
redis
mysqlclient
flask-cors
bcrypt
flask_sqlalchemy
mysql-connector
pyjwt
